fixed UpdateGroup usage

+++ src/parallelization/hiparsimulator/test/parallel_mpi_1/vanillastepperbasictest.h
@@ -52,10 +52,10 @@ public:
     void testPutPatch()
     {
         stepper->update(9);
-        TS_ASSERT_EQUALS(2, patchAccepter->offeredNanoSteps.size());
+        TS_ASSERT_EQUALS(size_t(2), patchAccepter->offeredNanoSteps.size());
 
         stepper->update(4);
-        TS_ASSERT_EQUALS(3, patchAccepter->offeredNanoSteps.size());
+        TS_ASSERT_EQUALS(size_t(3), patchAccepter->offeredNanoSteps.size());
     }
 
 private:
